feat(runner): carry environment variables with --env-file
