+++ src/core/messages/index.ts
@@ -4,7 +4,7 @@ export { environmentsMessages } from "./environments.js";
 export { flowsMessages } from "./flows.js";
 export { initMessages } from "./init.js";
 export { installMessages } from "./install.js";
-export { interactiveRunnerMessages } from "./interactiveRunner.js";
+export { interactiveRunnerMessages } from "./interactiveRunner/index.js";
 export { runnerMessages } from "./runner.js";
 export { packageLoadFailed } from "./toolNotFound.js";
 export { updateCheckMessages } from "./updateCheck.js";
