+++ src/core/messages/interactiveRunner/index.ts
@@ -1,0 +1,9 @@
+import { interactMessages } from "./interact.js";
+import { lifecycleMessages } from "./lifecycle.js";
+import { runMessages } from "./run.js";
+
+export const interactiveRunnerMessages = {
+  ...lifecycleMessages,
+  ...runMessages,
+  ...interactMessages,
+} as const;
